refactor: Remove useless native implementations
